+++ src/main/java/codesquad/http/handler/user/UserApiHandler.java
@@ -1,4 +1,4 @@
-package codesquad.http.handler;
+package codesquad.http.handler.user;
 
 import static codesquad.server.WebWorker.*;
