+++ src/main/java/codesquad/http/handler/user/UserHandler.java
@@ -1,4 +1,4 @@
-package codesquad.http.handler;
+package codesquad.http.handler.user;
 
 import static codesquad.session.SessionManager.*;
